Fixed serialization of GenderDetectStrategy, per PR recommendation to test with local workflow
